CORE-5400 /Refactoring/added thread to cleanup log tables
