+++ coordinator/src/main/scala/com/socrata/datacoordinator/service/Main.scala
@@ -105,12 +105,6 @@ class Main(common: SoQLCommon, serviceConfig: ServiceConfig) {
     }
   }
 
-  def removeDataset (datasetId: DatasetId) = {
-    for (u <- common.universe) yield {
-      u.datasetRemover.removeDataset(datasetId)
-    }
-  }
-
   def exporter(
     id: DatasetId,
     schemaHash: Option[String],
@@ -259,11 +253,34 @@ object Main {
 
         val serv = new Service(serviceConfig, operations.processMutation, operations.processCreation, getSchema _, getRollups,
           operations.exporter, secondaries, operations.datasetsInStore, operations.versionInStore,
-          operations.ensureInSecondary, operations.ensureInSecondaryGroup, operations.secondariesOfDataset, operations.listDatasets, operations.deleteDataset, operations.removeDataset,
+          operations.ensureInSecondary, operations.ensureInSecondaryGroup, operations.secondariesOfDataset, operations.listDatasets, operations.deleteDataset,
           serviceConfig.commandReadLimit, common.internalNameFromDatasetId, common.datasetIdFromInternalName, operations.makeReportTemporaryFile)
 
         val finished = new CountDownLatch(1)
+        val logTableCleanup = new Thread() {
+          setName("logTableCleanup thread")
+          override def run() {
+            do {
+              try {
+                for (u <- common.universe) {
+                  while (finished.getCount > 0 && u.logTableCleanup.cleanupOldVersions()) {
+                    u.commit()
+                    // a simple knob to allow us to slow the log table cleanup down without
+                    // requiring complicated things.
+                    finished.await(serviceConfig.logTableCleanupSleepTime.toMillis, TimeUnit.MILLISECONDS)
+                  }
+                }
+              } catch {
+                case e: Exception =>
+                  log.error("Unexpected error while cleaning log tables", e)
+              }
+            } while(!finished.await(30, TimeUnit.SECONDS))
+          }
+        }
+
+
         try {
+          logTableCleanup.start()
           val address = serviceConfig.discovery.address
           for {
             curator <- CuratorFromConfig(serviceConfig.curator)
