+++ coordinator/src/main/scala/com/socrata/datacoordinator/service/Service.scala
@@ -68,7 +68,6 @@ class Service(serviceConfig: ServiceConfig,
               secondariesOfDataset: DatasetId => Map[String, Long],
               datasets: () => Seq[DatasetId],
               deleteDataset: DatasetId => DatasetDropper.Result,
-              removeDataset: DatasetId => DatasetRemover.Result,
               commandReadLimit: Long,
               formatDatasetId: DatasetId => String,
               parseDatasetId: String => Option[DatasetId],
@@ -499,8 +498,7 @@ class Service(serviceConfig: ServiceConfig,
 
   case class DatasetResource(datasetId: DatasetId) extends SodaResource {
     override def post = doMutation
-    //override def delete = doDeleteDataset
-    override def delete = doRemoveDataset
+    override def delete = doDeleteDataset
     override def get = doExportFile
 
     val dateTimeFormat = ISODateTimeFormat.dateTime
@@ -569,22 +567,6 @@ class Service(serviceConfig: ServiceConfig,
       }
     }
 
-    def doRemoveDataset (req: HttpRequest): HttpResponse = {
-        removeDataset(datasetId) match {
-        case DatasetRemover.Success =>
-        OK ~>
-        Header("X-SODA2-Truth-Last-Modified", dateTimeFormat.print(DateTime.now)) ~>
-        Header("X-SODA2-Truth-Copy-Number", 0.toString) ~>
-        Header("X-SODA2-Truth-Version", 0.toString) ~>
-        Content(JsonContentType, "[]")
-        case DatasetRemover.FailureNotFound =>
-        notFoundError(formatDatasetId(datasetId))
-        case DatasetRemover.FailureWriteLock =>
-        writeLockError(datasetId)
-
-      }
-
-    }
 
     val suffixHashAlg = "SHA1"
     val suffixHashLen = MessageDigest.getInstance(suffixHashAlg).getDigestLength
