+++ coordinatorlib/src/main/scala/com/socrata/datacoordinator/truth/loader/sql/SqlTableDropper.scala
@@ -5,7 +5,7 @@ import java.sql.Connection
 
 class SqlTableDropper(conn: Connection) extends Closeable {
   //TODO: Store system_id in pending_table_drops so that we can figure out which metadata to delete with the dataset.
-  val stmt = conn.prepareStatement("INSERT INTO pending_table_drops (table_name, queued_at) values (?, now())")
+  val stmt = conn.createStatement()
 
   def close() {
     stmt.close()
@@ -15,9 +15,12 @@ class SqlTableDropper(conn: Connection) extends Closeable {
     stmt.executeBatch()
   }
 
-  def scheduleForDropping(tableName: String) {
-    stmt.setString(1, tableName)
-    stmt.addBatch()
+  def delete(tableName: String) {
+    stmt.executeUpdate("DROP TABLE IF EXISTS " + tableName)
+  }
+
+  def scheduleForDropping (tableName: String): Unit = {
+   stmt.executeUpdate("INSERT INTO pending_table_drops (table_name, queued_at) values (?, now())")
   }
 
 }
