+++ coordinatorlib/src/main/scala/com/socrata/datacoordinator/truth/universe/Universe.scala
@@ -119,10 +119,6 @@ trait DatasetDropperProvider {
   def datasetDropper: DatasetDropper
 }
 
-trait DatasetRemoverProvider {
-  def datasetRemover: DatasetRemover
-}
-
 trait TableCleanupProvider {
   def tableCleanup: TableCleanup
 }
