+++ coordinatorlib/src/main/scala/com/socrata/datacoordinator/truth/universe/sql/PostgresUniverse.scala
@@ -89,7 +89,6 @@ class PostgresUniverse[ColumnType, ColumnValue](conn: Connection,
     with DatasetReaderProvider
     with DatasetMutatorProvider
     with DatasetDropperProvider
-    with DatasetRemoverProvider
     with TableCleanupProvider
     with LogTableCleanupProvider
 {
@@ -190,8 +189,6 @@ class PostgresUniverse[ColumnType, ColumnValue](conn: Connection,
   lazy val datasetDropper =
     new SqlDatasetDropper(conn, writeLockTimeout, datasetMapWriter)
 
-  lazy val datasetRemover =
-    new SqlDatasetRemover(conn, writeLockTimeout, datasetMapWriter)
 
   lazy val tableCleanup: TableCleanup =
     new SqlTableCleanup(conn, tableCleanupDelay)
